+++ app/assets/javascripts/components/cards/Card.js
