+++ app/assets/javascripts/components/commons/Footer.js
