+++ app/assets/javascripts/components/commons/Map.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 
 class Map extends React.Component {
   componentDidMount() {
