+++ app/assets/javascripts/components/commons/Navbar.js
