+++ app/assets/javascripts/components/commons/Title.js
