+++ app/assets/javascripts/components/commons/ToolbarActions.js
