+++ app/assets/javascripts/components/commons/VegaChart.js
